Fix: After refreshing the codes, the tasks will regenerate
There was a problem where after refreshing the website the tasks wouldn't regenerate, instead the user would see what has been saved in local storage. I have fixed this bug with showTask function which checks the local storage and based on the data, it regenerate the tasks but there was also one problem, after each refresh there would be span added to the li that has "x" word in it. I fixed this bug with the using .nodeValue in the saveData function